Save file timestamps as they are on eridanus
I'm not planning to add new blog posts to this site. However, since I
will be leaving Uberspace, I'm saving the file timestamps as they are
right now on the host which has served my blog since 2013.
Unfortunately, Git alone won't help with that.

The idea is to restore the file mtimes during `docker build`.

This file was created using:

    ssh eridanus 'cd www/scytale.name/blog && ~/tsbackup.sh' \
        > timestamps.txt